+++ src/sem/bcknd/device/cuda/coef.cu
@@ -50,20 +50,18 @@ extern "C" {
                               void *jacinv, void *w3, int *nel,
                               int *lx, int *gdim) {
 
-    const dim3 nthrds(1024, 1, 1);
-    const dim3 nblcks((*nel), 1, 1);
     const cudaStream_t stream = (cudaStream_t) glb_cmd_queue;
 
 #define GEO_CASE(LX)                                                            \
     case LX:                                                                    \
-      coef_generate_geo_kernel<real, LX, 1024>                                  \
-        <<<nblcks, nthrds, 0, stream>>>                                         \
+      coef_generate_geo_kernel<real, LX, COEF_EB(LX)>                           \
+        <<<COEF_EB_NBLCKS(*nel, LX), COEF_EB_NTHRDS(LX), 0, stream>>>           \
         ((real *) G11, (real *) G12, (real *) G13,                              \
          (real *) G22, (real *) G23, (real *) G33,                              \
          (real *) drdx, (real *) drdy, (real *) drdz,                           \
          (real *) dsdx, (real *) dsdy, (real *) dsdz,                           \
          (real *) dtdx, (real *) dtdy, (real *) dtdz,                           \
-         (real *) jacinv, (real *) w3, *gdim);                                  \
+         (real *) jacinv, (real *) w3, *nel, *gdim);                            \
       CUDA_CHECK(cudaGetLastError());                                           \
       break
 
@@ -107,19 +105,18 @@ extern "C" {
 
     const int n = (*nel) * (*lx) * (*lx) * (*lx);
     const dim3 nthrds(1024, 1, 1);
-    const dim3 nblcks_dxyz((*nel), 1, 1);
     const dim3 nblcks_drst((n + 1024 - 1)/ 1024, 1, 1);
     const cudaStream_t stream = (cudaStream_t) glb_cmd_queue;
 
 #define DXYZDRST_CASE(LX)					               \
     case LX:								       \
-      coef_generate_dxyz_kernel<real, LX, 1024>                                \
-	<<<nblcks_dxyz, nthrds, 0, stream>>>                                   \
+      coef_generate_dxyz_kernel<real, LX, COEF_EB(LX)>                         \
+	<<<COEF_EB_NBLCKS(*nel, LX), COEF_EB_NTHRDS(LX), 0, stream>>>          \
         ((real *) dxdr, (real *) dydr, (real *) dzdr,                          \
          (real *) dxds, (real *) dyds, (real *) dzds,                          \
          (real *) dxdt, (real *) dydt, (real *) dzdt,                          \
          (real *) dx, (real *) dy, (real *) dz,                                \
-         (real *) x, (real *) y, (real *) z);                                  \
+         (real *) x, (real *) y, (real *) z, *nel);                            \
       CUDA_CHECK(cudaGetLastError());					       \
       break
 
@@ -190,19 +187,17 @@ extern "C" {
                                           void *wx, void *wy, void *wz,
                                           int *lx, int *nel, real eps) {
 
-    const dim3 nblcks((*nel), 1, 1);
-    const dim3 nthrds(1024, 1, 1);
     const cudaStream_t stream = (cudaStream_t) glb_cmd_queue;
 
 #define AREA_CASE(LX)                                                           \
     case LX:                                                                    \
-      coef_generate_area_and_normal_kernel<real, LX>                            \
-          <<<nblcks, nthrds, 0, stream>>>                                       \
+      coef_generate_area_and_normal_kernel<real, LX, COEF_EB(LX)>               \
+          <<<COEF_EB_NBLCKS(*nel, LX), COEF_EB_NTHRDS(LX), 0, stream>>>         \
           ((real *) area, (real *) nx, (real *) ny, (real *) nz,                \
            (real *) dxdr, (real *) dydr, (real *) dzdr,                         \
            (real *) dxds, (real *) dyds, (real *) dzds,                         \
            (real *) dxdt, (real *) dydt, (real *) dzdt,                         \
-           (real *) wx, (real *) wy, (real *) wz, eps);                         \
+           (real *) wx, (real *) wy, (real *) wz, eps, *nel);                   \
       CUDA_CHECK(cudaGetLastError());                                           \
       break